feedback - rename contants.ts to constants.ts
